jwt_auth/requirements.txt
asgiref==3.6.0
cffi==1.15.1
distlib==0.3.6
Django==4.1.7
django-dotenv==1.4.2
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
pbr==5.11.1
platformdirs==3.0.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
stevedore==5.0.0
tzdata==2022.7
virtualenv==20.19.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
